move `ToolSearch` advanced search jest to a new `ToolsList` jest
